requirements.txt
transformers
torch
pandas
scikit-learn
umap-learn
